Move LBF prefactors: nonbonded.py -> prefactor.py
Also remove unused switch_fn: nonbonded.py -> attic
