+++ attic/modules/potentials/switching_fns.py
@@ -1,0 +1,5 @@
+from jax import numpy as jnp
+
+
+def switch_fn(dij, cutoff):
+    return jnp.power(jnp.cos((jnp.pi * jnp.power(dij, 8)) / (2 * cutoff)), 2)
