+++ timemachine/potentials/nonbonded.py
@@ -1,30 +1,18 @@
-from typing import Any
+from typing import Any, Optional
 
 import jax.numpy as jnp
 import numpy as np
-from jax import jit, vmap
 from jax.scipy.special import erfc
 from numpy.typing import NDArray
-from scipy.special import binom
 
-from timemachine.potentials import jax_utils
 from timemachine.potentials.jax_utils import (
-    DEFAULT_CHUNK_SIZE,
     delta_r,
     distance_on_pairs,
     pairs_from_interaction_groups,
     pairwise_distances,
-    process_traj_in_chunks,
 )
 
 Array = Any
-
-
-def switch_fn(dij, cutoff):
-    return jnp.power(jnp.cos((jnp.pi * jnp.power(dij, 8)) / (2 * cutoff)), 2)
-
-
-from typing import Optional
 
 
 def combining_rule_sigma(sig_i, sig_j):
@@ -453,312 +441,3 @@ def validate_coulomb_cutoff(cutoff=1.0, beta=2.0, threshold=1e-2):
     following https://github.com/proteneer/timemachine/pull/424#discussion_r629678467"""
     if erfc(beta * cutoff) > threshold:
         print(UserWarning(f"erfc(beta * cutoff) = {erfc(beta * cutoff)} > threshold = {threshold}"))
-
-
-# utilities for efficiently recomputing energy as a function of ligand charges
-#   (where x_ligand, x_environment, q_environment are all constant, but q_ligand is variable)
-#   exploiting the fact that nonbonded_interaction_group(ligand_charges) is a linear function of ligand_charges
-#   TODO: avoid repetition between this and lennard-jones
-
-
-def coulomb_prefactor_on_atom(x_i, x_others, q_others, box=None, beta=2.0, cutoff=jnp.inf) -> float:
-    """Precompute part of (sum_i q_i * q_j / d_ij * rxn_field(d_ij)) that does not depend on q_i
-
-    Parameters
-    ----------
-    x_i : [D] array
-        position of focus atom (in ligand)
-    x_others: [N_env, D] array
-        positions of all other atoms (in environment)
-    q_others: [N_env] array
-        charges of all other atoms (in environment)
-    box: optional diagonal [D, D] array
-    beta: float
-    cutoff: float
-
-    Returns
-    -------
-    prefactor_i : float
-        sum_j q_j / d_ij * erfc(beta * d_ij)
-    """
-    d_ij = jax_utils.distance_from_one_to_others(x_i, x_others, box, cutoff)
-    prefactor_i = jnp.sum((q_others / d_ij) * erfc(beta * d_ij))
-    return prefactor_i
-
-
-def coulomb_prefactors_on_snapshot(x_ligand, x_env, q_env, box=None, beta=2.0, cutoff=np.inf) -> Array:
-    """Map coulomb_prefactor_on_atom over atoms in x_ligand
-
-    Parameters
-    ----------
-    x_ligand: [N_lig, D] array
-    x_env: [N_env, D] array
-    q_env: [N_env] array
-    box: optional diagonal [D, D] array
-    beta: float
-    cutoff: float
-
-    Returns
-    -------
-    prefactors: [N_lig] array
-        prefactors[i] = coulomb_prefactor_on_atom(x_ligand[i], ...)
-    """
-
-    def f_atom(x_i):
-        return coulomb_prefactor_on_atom(x_i, x_env, q_env, box, beta, cutoff)
-
-    return vmap(f_atom)(x_ligand)
-
-
-def coulomb_prefactors_on_traj(
-    traj, boxes, charges, ligand_indices, env_indices, beta=2.0, cutoff=jnp.inf, chunk_size=DEFAULT_CHUNK_SIZE
-):
-    """Map coulomb_prefactors_on_snapshot over snapshots in a trajectory
-
-    Parameters
-    ----------
-    traj: [T, N, D] array
-    boxes: diagonal [T, D, D] array (or [T] array of Nones)
-    charges: [N] array
-    ligand_indices: [N_lig] array of ints
-    env_indices: [N_env] array of ints
-    beta: float
-    cutoff: float
-    chunk_size: int
-        process traj in ceil(T / chunk_size) chunks, to limit memory consumption
-
-    Returns
-    -------
-    traj_prefactors: [T, N_lig] array
-        traj_prefactors[t] = coulomb_prefactors_on_snapshot(traj[t][ligand_indices], ...)
-    """
-    validate_interaction_group_idxs(len(traj[0]), ligand_indices, env_indices)
-
-    q_env = charges[env_indices]
-
-    def f_snapshot(coords, box):
-        x_ligand = coords[ligand_indices]
-        x_env = coords[env_indices]
-        return jit(coulomb_prefactors_on_snapshot)(x_ligand, x_env, q_env, box, beta, cutoff)
-
-    return process_traj_in_chunks(f_snapshot, traj, boxes, chunk_size)
-
-
-def coulomb_interaction_group_energy(q_ligand: Array, q_prefactors: Array) -> float:
-    """Assuming q_prefactors = coulomb_prefactors_on_snapshot(x_ligand, ...),
-    cheaply compute the energy of ligand-environment interaction group
-
-    Parameters
-    ----------
-    q_ligand: [N_lig] array
-    q_prefactors: [N_lig] array
-
-    Returns
-    -------
-    energy: float
-    """
-
-    return jnp.dot(q_prefactors, q_ligand)
-
-
-# utilities for efficiently recomputing energy as a function of ligand LJ parameters
-#   (where (x_ligand, x_env, sig_env, eps_env) are all constant, but (sig_ligand, eps_ligand) are variable)
-#   using [Naden, Shirts]'s linear basis-function approach
-#   TODO: avoid repetition between this and coulomb
-
-
-def _basis_expand_lj_term(sig_env, eps_env, r_env, power):
-    """
-    Compute expansion of
-        sum_i (4 * eps * eps_env[i] * (sig_env[i] / r_env[i])^power)
-
-    (called by basis_expand_lj_env with power=12 and power=6)
-
-    Parameters
-    ----------
-    sig_env, eps_env : [N_env] arrays
-        sigma, epsilon parameters of environment atoms
-    r_env : [N_env] array
-        distances from trial atom to all environment atoms
-    power : int
-        12 or 6
-
-    Returns
-    -------
-    h_n : [power + 1] array
-
-    References
-    ----------
-    eq. C.1 of Levi Naden's thesis
-    """
-    r_inv_pow = r_env**-power
-
-    exponents = power - np.arange(power + 1)
-    coeffs = binom(power, exponents)
-
-    raised = sig_env ** jnp.expand_dims(exponents, 1)
-    h_n_i = r_inv_pow * raised * jnp.expand_dims(coeffs, 1) * jnp.expand_dims(eps_env, 0)
-
-    h_n = jnp.sum(4 * h_n_i, 1)
-    return h_n
-
-
-def basis_expand_lj_env(sig_env, eps_env, r_env):
-    """Precomputed part of basis expansion that allows fast computation of
-
-    f(sig, eps)
-        = sum_i LJ(r_env[i]; sig + sig_env[i], eps * eps_env[i])
-        = dot(
-            basis_expand_lj_atom(sig, eps),
-            basis_expand_lj_env(sig_env, eps_env, r_env)
-        )
-    Parameters
-    ----------
-    sig_env, eps_env : [N_env] arrays
-        sigma, epsilon parameters of environment atoms
-    r_env : [N_env] array
-        distances from variable atom to all environment atoms
-
-    Returns
-    -------
-    h_n : [20] array
-        can be dotted with output of basis_expand_lj_atom(sig, eps)
-        to compute energy of one atom interacting with all environment atoms
-    """
-    jit_basis_expand_lj_term = jit(_basis_expand_lj_term, static_argnums=3)
-    h_n_12 = jit_basis_expand_lj_term(sig_env, eps_env, r_env, 12)
-    h_n_6 = -jit_basis_expand_lj_term(sig_env, eps_env, r_env, 6)
-    return jnp.hstack([h_n_12, h_n_6])
-
-
-def basis_expand_lj_atom(sig: float, eps: float) -> Array:
-    """Variable part of basis expansion that allows fast computation of
-
-    f(sig, eps)
-        = sum_i LJ(r_env[i]; sig + sig_env[i], eps * eps_env[i])
-        = dot(
-            basis_expand_lj_atom(sig, eps),
-            basis_expand_lj_env(sig_env, eps_env, r_env)
-        )
-
-    Parameters
-    ----------
-    sig, eps : floats
-        LJ parameters of variable atom
-
-    Returns
-    -------
-    projection : [20] array
-        can be dotted with output of basis_expand_lj_env(sig_env, eps_env, r_env)
-        to compute energy of one atom interacting with all environment atoms
-    """
-    exponents = jnp.hstack([jnp.arange(12 + 1), jnp.arange(6 + 1)])
-    return eps * (sig**exponents)
-
-
-def lj_prefactors_on_atom(x, x_others, sig_others, eps_others, box=None, cutoff=jnp.inf):
-    """Precompute part of
-
-        sum_j LennardJones(x_i, x_j; (sig_i, eps_i), (sig_j, eps_j))
-
-    that does not depend on (sig_i, eps_i)
-
-    Parameters
-    ----------
-    x : [D] array
-        position of focus atom (in ligand)
-    x_others: [N_env, D] array
-        positions of all other atoms (in environment)
-    sig_others, eps_others: [N_env] arrays
-        Lennard-Jones parameters of all other atoms (in environment)
-    box: optional diagonal [D, D] array
-    cutoff: float
-
-    Returns
-    -------
-    prefactors : [20] array
-
-    See Also
-    --------
-    basis_expand_lj_atom : computes a basis expansion of an atom's (sig, eps) that can be dotted with these prefactors
-    """
-    r_env = jax_utils.distance_from_one_to_others(x, x_others, box, cutoff)
-    prefactors = basis_expand_lj_env(sig_others, eps_others, r_env)
-    return prefactors
-
-
-def lj_prefactors_on_snapshot(x_ligand, x_env, sig_env, eps_env, box=None, cutoff=jnp.inf):
-    """Map lj_prefactor_on_atom over atoms in x_ligand
-
-    Parameters
-    ----------
-    x_ligand: [N_lig, D] array
-    x_env: [N_env, D] array
-    sig_env: [N_env] array
-    eps_env: [N_env] array
-    box: optional diagonal [D, D] array
-    cutoff: float
-
-    Returns
-    -------
-    prefactors: [N_lig, 20] array
-        prefactors[i] = lj_prefactor_on_atom(x_ligand[i], ...)
-    """
-
-    def f_atom(x_i):
-        return lj_prefactors_on_atom(x_i, x_env, sig_env, eps_env, box, cutoff)
-
-    return vmap(f_atom)(x_ligand)
-
-
-def lj_prefactors_on_traj(
-    traj, boxes, sigmas, epsilons, ligand_indices, env_indices, cutoff=jnp.inf, chunk_size=DEFAULT_CHUNK_SIZE
-):
-    """Map lj_prefactors_on_snapshot over snapshots in a trajectory
-
-    Parameters
-    ----------
-    traj: [T, N, D] array
-    boxes: diagonal [T, D, D] array (or [T] array of Nones)
-    sigmas: [N] array
-    epsilons: [N] array
-    ligand_indices: [N_lig] array of ints
-    env_indices: [N_env] array of ints
-    cutoff: float
-    chunk_size: int
-        process traj in ceil(T / chunk_size) chunks, to limit memory consumption
-
-    Returns
-    -------
-    traj_prefactors: [T, N_lig, 20] array
-        traj_prefactors[t] = lj_prefactors_on_snapshot(traj[t][ligand_indices], ...)
-    """
-    validate_interaction_group_idxs(len(traj[0]), ligand_indices, env_indices)
-
-    eps_env = epsilons[env_indices]
-    sig_env = sigmas[env_indices]
-
-    def f_snapshot(coords, box):
-        x_ligand = coords[ligand_indices]
-        x_env = coords[env_indices]
-        return jit(lj_prefactors_on_snapshot)(x_ligand, x_env, sig_env, eps_env, box, cutoff)
-
-    return process_traj_in_chunks(f_snapshot, traj, boxes, chunk_size)
-
-
-def lj_interaction_group_energy(sig_ligand, eps_ligand, lj_prefactors):
-    """Assuming lj_prefactors = lj_prefactors_on_snapshot(x_ligand, ...),
-    cheaply compute the energy of ligand-environment interaction group
-
-    Parameters
-    ----------
-    sig_ligand, eps_ligand: [N_lig] arrays
-    lj_prefactors: [N_lig, 20] array
-
-    Returns
-    -------
-    energy: float
-    """
-
-    projection = vmap(basis_expand_lj_atom)(sig_ligand, eps_ligand)
-    return jnp.sum(projection * lj_prefactors)
